move header releated components
