+++ src/sentry/static/sentry/app/components/events/interfaces/spans/dividerHandlerManager.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
+import {toPercent} from 'app/components/waterfallTree/utils';
 import {setBodyUserSelect, UserSelectValues} from 'app/utils/userselect';
 
-import {clamp, rectOfContent, toPercent} from './utils';
+import {clamp, rectOfContent} from './utils';
 
 // divider handle is positioned at 50% width from the left-hand side
 const DEFAULT_DIVIDER_POSITION = 0.4;
