+++ src/sentry/static/sentry/app/utils/discover/fieldRenderers.tsx
@@ -5,17 +5,17 @@ import partial from 'lodash/partial';
 
 import Count from 'app/components/count';
 import Duration from 'app/components/duration';
-import {
-  getHumanDuration,
-  pickSpanBarColour,
-  toPercent,
-} from 'app/components/events/interfaces/spans/utils';
+import {pickSpanBarColour} from 'app/components/events/interfaces/spans/utils';
 import ProjectBadge from 'app/components/idBadge/projectBadge';
 import UserBadge from 'app/components/idBadge/userBadge';
 import UserMisery from 'app/components/userMisery';
 import Version from 'app/components/version';
 import {DurationPill, RowRectangle} from 'app/components/waterfallTree/rowBar';
-import {getDurationDisplay} from 'app/components/waterfallTree/utils';
+import {
+  getDurationDisplay,
+  getHumanDuration,
+  toPercent,
+} from 'app/components/waterfallTree/utils';
 import {t} from 'app/locale';
 import {Organization} from 'app/types';
 import {defined} from 'app/utils';
