+++ src/sentry/static/sentry/app/views/performance/traceDetails/styles.tsx
@@ -16,13 +16,6 @@ import {appendTagCondition} from 'app/utils/queryString';
 import {Theme} from 'app/utils/theme';
 import {transactionSummaryRouteWithQuery} from 'app/views/performance/transactionSummary/utils';
 
-export {
-  DividerSpacer,
-  ScrollBarContainer as ScrollbarContainer,
-  VirtualScrollBar,
-  VirtualScrollBarGrip,
-} from 'app/components/events/interfaces/spans/header';
-
 export const SearchContainer = styled('div')`
   display: flex;
   width: 100%;
